[Feat]: 좌석 예약 결제 처리

+++ src/main/java/io/dev/concertreservationsystem/application/pointHistory/PointHistoryAdminDTOParam.java
@@ -25,7 +25,7 @@ public record PointHistoryAdminDTOParam(
         @NotNull
         @NotBlank
         @Min(0)
-        Long amount
+        Integer amount
 ) {
 
     public PointHistoryDTOParam convertToPointHistoryDTOParam() {
