+++ src/main/java/io/dev/concertreservationsystem/application/pointHistory/PointHistoryAdminDTOResult.java
@@ -12,7 +12,7 @@ public record PointHistoryAdminDTOResult(
         Long paymentId,
         String userId,
         PointTransactionType type,
-        Long amount,
+        Integer amount,
         Long resultPoint,
         LocalDateTime created_at,
         LocalDateTime updated_at
