+++ src/main/java/io/dev/concertreservationsystem/domain/pointHistory/factory/SimplePointHistoryFactory.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Component;
 @Component
 public class SimplePointHistoryFactory extends PointHistoryFactory {
     @Override
-    public PointHistory createPointHistory(String userId, PointTransactionType type, Long amount, Long resultPoint){
+    public PointHistory createPointHistory(String userId, PointTransactionType type, Integer amount, Long resultPoint){
         return PointHistory.builder()
                 .userId(userId)
                 .type(type)
